adjusting some cython data types to match #41

+++ colibricore_classes.in.pxd
@@ -33,7 +33,7 @@ cdef extern from "pattern.h":
     cdef cppclass Pattern:
         unsigned char * data
         Pattern() nogil
-        Pattern(Pattern&, unsigned int,unsigned int) nogil
+        Pattern(Pattern&, long unsigned int,long unsigned int) nogil
         Pattern(Pattern&) nogil
         #Pattern(PatternPointer&) nogil
         #Pattern(PatternPointer&, int,int) nogil
@@ -69,14 +69,14 @@ cdef extern from "pattern.h":
 
     cdef cppclass PatternPointer:
         PatternPointer() nogil
-        PatternPointer(Pattern&, unsigned int,unsigned int) nogil
+        PatternPointer(Pattern&, long unsigned int,long unsigned int) nogil
         PatternPointer(Pattern&) nogil
-        PatternPointer(PatternPointer&, unsigned int,unsigned int) nogil
+        PatternPointer(PatternPointer&, long unsigned int,long unsigned int) nogil
         PatternPointer(PatternPointer&) nogil
         Pattern pattern() nogil
         string tostring(ClassDecoder&) nogil
-        unsigned int n() nogil
-        unsigned int bytesize() nogil
+        long unsigned int n() nogil
+        long unsigned int bytesize() nogil
         int skipcount() nogil
         int category() nogil
         int hash() nogil
